Updates to About
Added the basics for directing to club pages on GitHub and Slack, as
well as the club email. Also added Dr. Mike to the officers list, setup
the officers in a collapsing div (same as with Resources, though I might
not keep it), and made a couple changes so adding more officers should
be easy.

Also edited the nav-bar .css a tad (added links that move when clicked)